Move usages of @prisma/migrate and @prisma/sdk from adapter to @keystone-next/keystone
