+++ .changeset/dirty-moons-drive.md
@@ -1,0 +1,5 @@
+---
+'@keystone-next/adapter-prisma-legacy': major
+---
+
+Removed `CLIOptionsForCreateMigration` and `createMigration` exports
