--- .changeset/giant-buses-sip.md
@@ -1,5 +1,0 @@
----
-'@keystone-next/adapter-prisma-legacy': minor
----
-
-Added `devMigrations` and `runPrototypeMigrations` exports.
