+++ .changeset/ten-jokes-clean.md
@@ -1,0 +1,5 @@
+---
+'@keystone-next/keystone': minor
+---
+
+Added `migrations` entrypoint with `pushPrismaSchemaToDatabase` export.
