+++ packages-next/keystone/package.json
@@ -21,7 +21,10 @@
     "@keystone-next/keystone-legacy": "^22.0.0",
     "@keystone-next/server-side-graphql-client-legacy": "3.0.0",
     "@keystone-next/types": "^15.0.1",
+    "@prisma/client": "2.19.0",
+    "@prisma/migrate": "2.19.0",
     "@prisma/sdk": "2.19.0",
+    "@sindresorhus/slugify": "^1.1.0",
     "@types/babel__core": "^7.1.14",
     "@types/cookie": "^0.4.0",
     "@types/express": "^4.17.11",
@@ -36,6 +39,7 @@
     "apollo-server-express": "^2.22.2",
     "apollo-server-micro": "^2.22.2",
     "apollo-server-types": "^0.7.0",
+    "chalk": "^4.1.0",
     "cookie": "^0.4.1",
     "cors": "^2.8.5",
     "execa": "^5.0.0",
@@ -64,6 +68,7 @@
     "entrypoints": [
       "index.ts",
       "artifacts.ts",
+      "migrations.ts",
       "schema/index.ts",
       "session/index.ts",
       "scripts/index.ts"
