+++ packages-next/keystone/src/migrations.ts
@@ -1,0 +1,1 @@
+export { pushPrismaSchemaToDatabase } from './lib/migrations';
