+++ packages/adapter-prisma/src/index.ts
@@ -1,3 +1,1 @@
 export { PrismaAdapter, PrismaListAdapter, PrismaFieldAdapter } from './adapter-prisma';
-export { createMigration, devMigrations, runPrototypeMigrations } from './migrations';
-export type { CLIOptionsForCreateMigration } from './migrations';
